Do not perform venv clean install symlink check on Windows

+++ tests/meltano/core/test_venv_service.py
@@ -117,17 +117,18 @@ class TestVenvService:
         # Make sure the venv exists already
         await subject.install(["example"], clean=True)
 
-        original_link_target = os.readlink(subject.python_path)
-        try:
-            # Simulate the deletion of the underlying Python executable by
-            # making its symlink point to a file that does not exist
-            subject.python_path.unlink()
-            subject.python_path.symlink_to("./fake/path/to/python/executable")
-            assert subject.requires_clean_install(["example"])
-        finally:
-            # Restore the symlink to the actual Python executable
-            subject.python_path.unlink()
-            subject.python_path.symlink_to(original_link_target)
+        if platform.system() != "Windows":
+            original_link_target = os.readlink(subject.python_path)
+            try:
+                # Simulate the deletion of the underlying Python executable by
+                # making its symlink point to a file that does not exist
+                subject.python_path.unlink()
+                subject.python_path.symlink_to("./fake/path/to/python/executable")
+                assert subject.requires_clean_install(["example"])
+            finally:
+                # Restore the symlink to the actual Python executable
+                subject.python_path.unlink()
+                subject.python_path.symlink_to(original_link_target)
 
         assert not subject.requires_clean_install(["example"])
         assert subject.requires_clean_install(["example==0.1.0"])
